+++ ntex/src/web/responder.rs
@@ -34,9 +34,6 @@ impl<T> Future for Ready<T> {
 ///
 /// Types that implement this trait can be used as the return type of a handler.
 pub trait Responder<Err = DefaultError> {
-    /// The associated error which can be returned.
-    type Error;
-
     /// The future response value.
     type Future: Future<Output = Response>;
 
@@ -93,7 +90,6 @@ pub trait Responder<Err = DefaultError> {
 }
 
 impl<Err: ErrorRenderer> Responder<Err> for Response {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     #[inline]
@@ -103,7 +99,6 @@ impl<Err: ErrorRenderer> Responder<Err> for Response {
 }
 
 impl<Err: ErrorRenderer> Responder<Err> for ResponseBuilder {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     #[inline]
@@ -117,7 +112,6 @@ where
     T: Responder<Err>,
     Err: ErrorRenderer,
 {
-    type Error = T::Error;
     type Future = Either<T::Future, Ready<Response>>;
 
     fn respond_to(self, req: &HttpRequest) -> Self::Future {
@@ -136,7 +130,6 @@ where
     E: Into<Err::Container>,
     Err: ErrorRenderer,
 {
-    type Error = T::Error;
     type Future = Either<T::Future, Ready<Response>>;
 
     fn respond_to(self, req: &HttpRequest) -> Self::Future {
@@ -152,7 +145,6 @@ where
     T: Responder<Err>,
     Err: ErrorRenderer,
 {
-    type Error = T::Error;
     type Future = CustomResponderFut<T, Err>;
 
     fn respond_to(self, req: &HttpRequest) -> Self::Future {
@@ -165,7 +157,6 @@ where
 }
 
 impl<Err: ErrorRenderer> Responder<Err> for &'static str {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     fn respond_to(self, _: &HttpRequest) -> Self::Future {
@@ -178,7 +169,6 @@ impl<Err: ErrorRenderer> Responder<Err> for &'static str {
 }
 
 impl<Err: ErrorRenderer> Responder<Err> for &'static [u8] {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     fn respond_to(self, _: &HttpRequest) -> Self::Future {
@@ -191,7 +181,6 @@ impl<Err: ErrorRenderer> Responder<Err> for &'static [u8] {
 }
 
 impl<Err: ErrorRenderer> Responder<Err> for String {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     fn respond_to(self, _: &HttpRequest) -> Self::Future {
@@ -204,7 +193,6 @@ impl<Err: ErrorRenderer> Responder<Err> for String {
 }
 
 impl<'a, Err: ErrorRenderer> Responder<Err> for &'a String {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     fn respond_to(self, _: &HttpRequest) -> Self::Future {
@@ -217,7 +205,6 @@ impl<'a, Err: ErrorRenderer> Responder<Err> for &'a String {
 }
 
 impl<Err: ErrorRenderer> Responder<Err> for Bytes {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     fn respond_to(self, _: &HttpRequest) -> Self::Future {
@@ -230,7 +217,6 @@ impl<Err: ErrorRenderer> Responder<Err> for Bytes {
 }
 
 impl<Err: ErrorRenderer> Responder<Err> for BytesMut {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     fn respond_to(self, _: &HttpRequest) -> Self::Future {
@@ -322,7 +308,6 @@ impl<T: Responder<Err>, Err> CustomResponder<T, Err> {
 }
 
 impl<T: Responder<Err>, Err: ErrorRenderer> Responder<Err> for CustomResponder<T, Err> {
-    type Error = T::Error;
     type Future = CustomResponderFut<T, Err>;
 
     fn respond_to(self, req: &HttpRequest) -> Self::Future {
@@ -390,7 +375,6 @@ where
     B: Responder<Err>,
     Err: ErrorRenderer,
 {
-    type Error = Err::Container;
     type Future = Either<A::Future, B::Future>;
 
     fn respond_to(self, req: &HttpRequest) -> Self::Future {
@@ -406,7 +390,6 @@ where
     T: std::fmt::Debug + std::fmt::Display + 'static,
     Err: ErrorRenderer,
 {
-    type Error = Err::Container;
     type Future = Ready<Response>;
 
     fn respond_to(self, req: &HttpRequest) -> Self::Future {
@@ -423,9 +406,7 @@ pub(crate) mod tests {
     use crate::web::test::{init_service, TestRequest};
     use crate::{service::Service, util::Bytes, util::BytesMut, web};
 
-    fn responder<T: Responder<DefaultError>>(
-        responder: T,
-    ) -> impl Responder<DefaultError, Error = T::Error> {
+    fn responder<T: Responder<DefaultError>>(responder: T) -> impl Responder<DefaultError> {
         responder
     }